+++ changelog.md
@@ -1,4 +1,6 @@
-
 V 0.4.9
 
 - [20240501] Integrated changes "from KBS-aurious"
+  - Added *PatchFile* to Apply a patch in unified diff format.
+  - Added *Install-License*
+  - PKG *vfs* updated to version 1.4.2